* master: (63 commits)
  Move the Mistral config doc to the install section
  Minor fix to string substitution in get_action_execution_id_from_url
  Use warn instead of exception in between retries
  Fix retry logic in the Mistral callback handler
  Create get_action_execution_id_from_url method for Mistral callback
  Move Mistral unit tests for auth to a different file
  Update change log for configuring Mistral auth
  Add docs to configure auth options for Mistral
  Add config options for Mistral auth with Keystone
  Refactor Mistral callback handler to use python client instead of API
  Refactor Mistral querier to use python client instead of API
  Fix lint violations in the linux pack.
  Also run flake8 and pylint on Python files inside the linux pack.
  Improve logging in wait_for_ssh action, remove unecessary second sleep.
  Remove unused imports.
  wait_for_ssh should throw on timeout, otherwise the action will never fail unless the user checks / compares the result value.
  Make sure that the "default" pack which we create during installation has a pack.yaml file.
  Paramiko to be default SSH runner
  Add some missing license headers.
  Update st2-register-content CLI script to use the new script setup code.
  ...

+++ st2tests/st2tests/base.py
@@ -19,6 +19,7 @@ except ImportError:
     import json
 
 import os
+import os.path
 import sys
 import shutil
 
@@ -29,6 +30,8 @@ from unittest2 import TestCase
 
 from st2common.exceptions.db import StackStormDBObjectConflictError
 from st2common.models.db import db_setup, db_teardown
+from st2common.bootstrap.base import ResourceRegistrar
+from st2common.content.utils import get_packs_base_paths
 import st2common.models.db.rule as rule_model
 import st2common.models.db.sensor as sensor_model
 import st2common.models.db.trigger as trigger_model
@@ -131,6 +134,14 @@ class BaseDbTestCase(TestCase):
         for model in ALL_MODELS:
             model.drop_collection()
 
+    @classmethod
+    def _register_packs(self):
+        """
+        Register all the packs inside the fixtures directory.
+        """
+        registrar = ResourceRegistrar(use_pack_cache=False)
+        registrar.register_packs(base_dirs=get_packs_base_paths())
+
 
 class DbTestCase(BaseDbTestCase):
     """
@@ -142,11 +153,15 @@ class DbTestCase(BaseDbTestCase):
 
     db_connection = None
     current_result = None
+    register_packs = False
 
     @classmethod
     def setUpClass(cls):
         BaseDbTestCase.setUpClass()
         cls._establish_connection_and_re_create_db()
+
+        if cls.register_packs:
+            cls._register_packs()
 
     @classmethod
     def tearDownClass(cls):
@@ -255,19 +270,33 @@ class CleanDbTestCase(BaseDbTestCase):
 
 class CleanFilesTestCase(TestCase):
     """
-    Base test class which deletes specified files and directories on tearDown.
+    Base test class which deletes specified files and directories on setUp and `tearDown.
     """
     to_delete_files = []
     to_delete_directories = []
 
+    def setUp(self):
+        super(CleanFilesTestCase, self).setUp()
+        self._delete_files()
+
     def tearDown(self):
+        super(CleanFilesTestCase, self).tearDown()
+        self._delete_files()
+
+    def _delete_files(self):
         for file_path in self.to_delete_files:
+            if not os.path.isfile(file_path):
+                continue
+
             try:
                 os.remove(file_path)
             except Exception:
                 pass
 
         for file_path in self.to_delete_directories:
+            if not os.path.isdir(file_path):
+                continue
+
             try:
                 shutil.rmtree(file_path)
             except Exception:
